+++ packages/unpublished-relationship-performance-test-app/app/models/child.js
@@ -4,8 +4,8 @@ const { Model, attr, belongsTo, hasMany } = DS;
 
 export default Model.extend({
   childName: attr('string'),
-  friends: hasMany('child', { async: true }),
-  bestFriend: belongsTo('child', { async: true }),
-  secondBestFriend: belongsTo('child', { async: true }),
+  friends: hasMany('child', { async: true, inverse: null }),
+  bestFriend: belongsTo('child', { async: true, inverse: 'bestFriend' }),
+  secondBestFriend: belongsTo('child', { async: true, inverse: 'secondBestFriend' }),
   parent: belongsTo('parent', { async: true }),
 });
