+++ packages/unpublished-relationship-performance-test-app/app/routes/unload.js
@@ -1,43 +1,23 @@
 import Route from '@ember/routing/route';
-import { schedule } from '@ember/runloop';
-
-import { Promise } from 'rsvp';
+import { run } from '@ember/runloop';
 
 import { createParentPayload } from '../utils/create-parent-payload';
-import { endTrace } from '../utils/end-trace';
 
 export default Route.extend({
   model() {
     performance.mark('start-data-generation');
-
     const payload = createParentPayload(2000, 500);
-
-    performance.mark('end-data-generation');
-    performance.measure('data-generation', 'start-data-generation', 'end-data-generation');
-
     performance.mark('start-push-payload');
     const result = this.store.push(payload);
-    performance.mark('end-push-payload');
-    performance.measure('push-payload', 'start-push-payload', 'end-push-payload');
-
-    performance.mark('start-unloadRecord');
+    performance.mark('start-unload-records');
     const parent = result[0];
-    parent
-      .get('children')
-      .toArray()
-      .forEach((child) => child.unloadRecord());
-    parent.unloadRecord();
-
-    return new Promise((resolve, reject) => {
-      schedule('destroy', this, () => {
-        performance.mark('end-unloadRecord');
-        performance.measure('unloadRecord', 'start-unloadRecord', 'end-unloadRecord');
-        resolve();
-      });
+    run(() => {
+      parent
+        .get('children')
+        .toArray()
+        .forEach((child) => child.unloadRecord());
+      parent.unloadRecord();
     });
-  },
-
-  afterModel() {
-    endTrace();
+    performance.mark('end-unload-records');
   },
 });
