--- packages/unpublished-relationship-performance-test-app/app/utils/end-trace.js
@@ -1,11 +1,0 @@
-export function endTrace() {
-  if (document.location.href.indexOf('?tracing') !== -1 || document.location.href.indexOf('?tracerbench=true') !== -1) {
-    // just before paint
-    requestAnimationFrame(() => {
-      // after paint
-      requestAnimationFrame(() => {
-        document.location.href = 'about:blank';
-      });
-    });
-  }
-}
